house_pr/pyproject.toml
[project]
name = "house-pr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "scikit-learn==1.6.1",
    "streamlit>=1.48.1",
]
